Refactor ProductCard into separate file and add price range filter
- Moved ProductCard component to its own file for better organization
- Implemented price range filter functionality for products
  Users can now filter products based on a specified price range
