+++ Android/app/src/main/java/hr/foi/air/giveaway/MainActivity.kt
@@ -73,7 +73,6 @@ class MainActivity : ComponentActivity() {
                         }
                         composable("home") {
                             // HomePage()
-                            // ProductsCard(DummyProducts.generateMockProducts().first())
                             ProductsPage()
                         }
                     }
