+++ Android/app/src/main/java/hr/foi/air/giveaway/navigation/components/products/ProductsPage.kt
@@ -1,25 +1,20 @@
 package hr.foi.air.giveaway.navigation.components.products
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.Button
 import androidx.compose.material.ButtonDefaults
-import androidx.compose.material.Card
 import androidx.compose.material.Checkbox
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.OutlinedTextField
@@ -33,74 +28,16 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import hr.foi.air.giveaway.mockdataprodutc.MockProducts
-import hr.foi.air.giveaway.mockdataprodutc.Product
 import hr.foi.air.giveaway.ui.theme.AppTheme
 import hr.foi.air.giveaway.viewmodels.ProductsViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
 import hr.foi.air.giveaway.mockdataprodutc.ProductType
 
-@Composable
-fun ProductsCard(product: Product, onCartButtonClick: () -> Unit) {
-    Card(
-        modifier = Modifier
-            .padding(8.dp)
-            .width(IntrinsicSize.Max)
-            .height(IntrinsicSize.Max),
-        shape = MaterialTheme.shapes.medium
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(8.dp)
-        ) {
-            Image(
-                painter = painterResource(id = product.imageRes),
-                contentDescription = null,
-                modifier = Modifier
-                    .size(160.dp)
-                    .fillMaxWidth(),
-                contentScale = ContentScale.Crop
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = product.name,
-                style = MaterialTheme.typography.subtitle1,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis
-            )
-            Text(
-                text = product.description,
-                style = MaterialTheme.typography.body2,
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis
-            )
-            Spacer(modifier = Modifier.weight(1f))
-            Text(
-                text = "$${product.price}",
-                style = MaterialTheme.typography.subtitle2,
-                color = MaterialTheme.colors.primaryVariant
-            )
-            Button(
-                onClick = onCartButtonClick,
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(16.dp),
-                colors = ButtonDefaults.buttonColors(backgroundColor = MaterialTheme.colors.primaryVariant)
-            ) {
-                Text("Add to Cart", color = Color.White)
-            }
-        }
-    }
-}
+
 
 @Composable
 fun ProductsPage(
@@ -141,20 +78,41 @@ fun ProductsPage(
                     Text("Filters")
                 }
             } else {
-
-                OutlinedTextField(
-                    value = viewModel.priceFilter.value.toString(),
-                    onValueChange = { newValue ->
-                        viewModel.priceFilter.value = newValue.toDoubleOrNull() ?: 0.0
-                    },
-                    label = { Text("Filter by Price") },
-                    keyboardOptions = KeyboardOptions.Default.copy(
-                        keyboardType = KeyboardType.Number
-                    ),
+                Row(
                     modifier = Modifier
                         .fillMaxWidth()
-                        .padding(6.dp)
-                )
+                        .padding(6.dp),
+                    horizontalArrangement = Arrangement.SpaceBetween
+                ) {
+                    OutlinedTextField(
+                        value = viewModel.minPriceFilter.value.toString(),
+                        onValueChange = { newValue ->
+                            viewModel.minPriceFilter.value = newValue.toDoubleOrNull() ?: 0.0
+                        },
+                        label = { Text("Min Price") },
+                        keyboardOptions = KeyboardOptions.Default.copy(
+                            keyboardType = KeyboardType.Number
+                        ),
+                        modifier = Modifier
+                            .weight(1f)
+                            .padding(end = 4.dp)
+                    )
+                    OutlinedTextField(
+                        value = viewModel.maxPriceFilter.value.toString(),
+                        onValueChange = { newValue ->
+                            viewModel.maxPriceFilter.value = newValue.toDoubleOrNull() ?: 0.0
+                        },
+                        label = { Text("Max Price") },
+                        keyboardOptions = KeyboardOptions.Default.copy(
+                            keyboardType = KeyboardType.Number
+                        ),
+                        modifier = Modifier
+                            .weight(1f)
+                            .padding(end = 4.dp)
+                    )
+                }
+
+
 
                 LazyRow {
                     items(ProductType.values()) { productType ->
@@ -201,20 +159,12 @@ fun ProductsPage(
                         horizontalArrangement = Arrangement.Center
                     ) {
                         for (product in rowProducts) {
-                            ProductsCard(product, {})
+                            ProductCard(product, {})
                         }
                     }
                 }
             }
         }
-    }
-}
-
-@Preview
-@Composable
-fun ProductsCardPreview() {
-    AppTheme {
-        ProductsCard(MockProducts.generateMockProducts().first(), {})
     }
 }
